feat: web3 helpers
