+++ src/lib/solana.ts
@@ -107,30 +107,3 @@ export function getPlatformToolsVersions(): PlatformToolsVersions {
 
   return tools;
 }
-
-/** Solana cluster moniker */
-export type SolanaClusterMoniker =
-  | "mainnet-beta"
-  | "devnet"
-  | "testnet"
-  | "localnet";
-
-/**
- * Get a public Solana RPC endpoint for a cluster based on its moniker
- *
- * Note: These RPC URLs are rate limited and not suitable for production applications.
- */
-export function getPublicSolanaRpcUrl(cluster: SolanaClusterMoniker): string {
-  switch (cluster) {
-    case "devnet":
-      return "https://api.devnet.solana.com";
-    case "testnet":
-      return "https://api.testnet.solana.com";
-    case "mainnet-beta":
-      return "https://api.mainnet-beta.solana.com";
-    case "localnet":
-      return "http://127.0.0.1:8899";
-    default:
-      throw new Error("Invalid cluster moniker");
-  }
-}
